:truck: Move CPPMonoBehaviour class to the components directory
